requirements.txt
python-dotenv==1.0.1
pandas==2.2.3
numpy==2.2.4
SQLAlchemy==2.0.39
streamlit==1.43.2
plotly==6.0.1
pydantic==2.10.6
pydantic-settings==2.8.1

